--- aero_description/typeF/README.md
@@ -1,15 +1,0 @@
-# Aero Robot Setting
-
-## type FCDT
-
-- upper
-  -  typeF_upperbody
-
-- lifter
-  - typeC_lifter
-
-- wheel
-  - typeD_wheel
-
-- hand
-  - trx_s(T)
