--- aero_description/typeF/aero_upper.txt
@@ -1,28 +1,0 @@
-waist_y_joint
-waist_p_joint
-waist_r_joint
-l_shoulder_p_joint
-l_shoulder_r_joint
-l_shoulder_y_joint
-l_elbow_joint
-l_wrist_y_joint
-l_wrist_p_joint
-l_wrist_r_joint
-l_indexbase_joint
-l_indexmid_joint
-l_indexend_joint
-l_thumb_joint
-neck_y_joint
-neck_p_joint
-neck_r_joint
-r_shoulder_p_joint
-r_shoulder_r_joint
-r_shoulder_y_joint
-r_elbow_joint
-r_wrist_y_joint
-r_wrist_p_joint
-r_wrist_r_joint
-r_indexbase_joint
-r_indexmid_joint
-r_indexend_joint
-r_thumb_joint
