--- aero_description/typeF/moveit_config/config/joint_limits.yaml
@@ -1,164 +1,0 @@
-# joint_limits.yaml allows the dynamics properties specified in the URDF to be overwritten or augmented as needed
-# Specific joint properties can be changed with the keys [max_position, min_position, max_velocity, max_acceleration]
-# Joint limits can be turned off with [has_velocity_limits, has_acceleration_limits]
-joint_limits:
-  ankle_joint:
-    has_velocity_limits: true
-    max_velocity: 1.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  ankle_joint_mimic:
-    has_velocity_limits: false
-    max_velocity: 0
-    has_acceleration_limits: false
-    max_acceleration: 0
-  knee_joint:
-    has_velocity_limits: true
-    max_velocity: 1.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  knee_joint_mimic:
-    has_velocity_limits: false
-    max_velocity: 0
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_elbow_joint:
-    has_velocity_limits: true
-    max_velocity: 1
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_indexbase_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_indexend_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_indexmid_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_shoulder_p_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_shoulder_r_joint:
-    has_velocity_limits: true
-    max_velocity: 0.6
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_shoulder_y_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_thumb_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_wrist_p_joint:
-    has_velocity_limits: true
-    max_velocity: 1.3
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_wrist_r_joint:
-    has_velocity_limits: true
-    max_velocity: 1.3
-    has_acceleration_limits: false
-    max_acceleration: 0
-  l_wrist_y_joint:
-    has_velocity_limits: true
-    max_velocity: 1.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  neck_p_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  neck_r_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  neck_y_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_elbow_joint:
-    has_velocity_limits: true
-    max_velocity: 1
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_indexbase_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_indexend_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_indexmid_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_shoulder_p_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_shoulder_r_joint:
-    has_velocity_limits: true
-    max_velocity: 0.6
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_shoulder_y_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_thumb_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_wrist_p_joint:
-    has_velocity_limits: true
-    max_velocity: 1.3
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_wrist_r_joint:
-    has_velocity_limits: true
-    max_velocity: 1.3
-    has_acceleration_limits: false
-    max_acceleration: 0
-  r_wrist_y_joint:
-    has_velocity_limits: true
-    max_velocity: 1.5
-    has_acceleration_limits: false
-    max_acceleration: 0
-  waist_p_joint:
-    has_velocity_limits: true
-    max_velocity: 0.3
-    has_acceleration_limits: false
-    max_acceleration: 0
-  waist_r_joint:
-    has_velocity_limits: true
-    max_velocity: 0.3
-    has_acceleration_limits: false
-    max_acceleration: 0
-  waist_y_joint:
-    has_velocity_limits: true
-    max_velocity: 0.5
-    has_acceleration_limits: false
-    max_acceleration: 0
