--- aero_description/typeF/robot.cfg
@@ -1,6 +1,0 @@
-: upper
-aero_shop/typeF_upperbody
-aero_shop/trx_s
-: lower
-aero_shop/typeC_lifter
-aero_shop/typeE_wheel -> aero_move_base
