--- aero_description/typeF/robots/gazebo_controller_settings.yaml
@@ -1,90 +1,0 @@
-# Publish all joint states -----------------------------------
-joint_state_controller:
-  type: joint_state_controller/JointStateController
-  publish_rate: 50
-# Joint Trajectory Controllers -------------------------------
-head_controller:
-#  type: effort_controllers/JointTrajectoryController
-  type: velocity_controllers/JointTrajectoryController
-  joints:
-    - neck_y_joint
-    - neck_p_joint
-    - neck_r_joint
-  gains:
-    neck_y_joint: {p: 40.0, i: 0.0, d: 0.0}
-    neck_p_joint: {p: 40.0, i: 0.0, d: 0.0}
-    neck_r_joint: {p: 40.0, i: 0.0, d: 0.0}
-waist_controller:
-#  type: effort_controllers/JointTrajectoryController
-  type: velocity_controllers/JointTrajectoryController
-  joints:
-    - waist_y_joint
-    - waist_p_joint
-    - waist_r_joint
-  gains:
-    waist_y_joint: {p: 100.0, i: 0.0, d: 0.0}
-    waist_p_joint: {p: 200.0, i: 0.0, d: 0.0}
-    waist_r_joint: {p: 200.0, i: 0.0, d: 0.0}
-rarm_controller:
-#  type: effort_controllers/JointTrajectoryController
-  type: velocity_controllers/JointTrajectoryController
-  joints:
-    - r_shoulder_p_joint
-    - r_shoulder_r_joint
-    - r_shoulder_y_joint
-    - r_elbow_joint
-    - r_wrist_y_joint
-    - r_wrist_p_joint
-    - r_wrist_r_joint
-  gains:
-    r_shoulder_p_joint: {p:  200.0, i: 0.0, d: 0.0}
-    r_shoulder_r_joint: {p:  200.0, i: 0.0, d: 0.0}
-    r_shoulder_y_joint: {p:  100.0, i: 0.0, d: 0.0}
-    r_elbow_joint:      {p:  100.0, i: 0.0, d: 0.0}
-    r_wrist_y_joint:    {p:  100.0, i: 0.0, d: 0.0}
-    r_wrist_p_joint:    {p:  100.0, i: 0.0, d: 0.0}
-    r_wrist_r_joint:    {p:  100.0, i: 0.0, d: 0.0}
-larm_controller:
-#  type: effort_controllers/JointTrajectoryController
-  type: velocity_controllers/JointTrajectoryController
-  joints:
-    - l_shoulder_p_joint
-    - l_shoulder_r_joint
-    - l_shoulder_y_joint
-    - l_elbow_joint
-    - l_wrist_y_joint
-    - l_wrist_p_joint
-    - l_wrist_r_joint
-  gains:
-    l_shoulder_p_joint: {p:  200.0, i: 0.0, d: 0.0}
-    l_shoulder_r_joint: {p:  200.0, i: 0.0, d: 0.0}
-    l_shoulder_y_joint: {p:  100.0, i: 0.0, d: 0.0}
-    l_elbow_joint:      {p:  100.0, i: 0.0, d: 0.0}
-    l_wrist_y_joint:    {p:  100.0, i: 0.0, d: 0.0}
-    l_wrist_p_joint:    {p:  100.0, i: 0.0, d: 0.0}
-    l_wrist_r_joint:    {p:  100.0, i: 0.0, d: 0.0}
-lifter_controller:
-#  type: effort_controllers/JointTrajectoryController
-  type: velocity_controllers/JointTrajectoryController
-  joints:
-    - knee_joint
-    - ankle_joint
-  gains:
-    knee_joint:  {p: 300.0, i: 0.0, d: 0.0}
-    ankle_joint: {p: 300.0, i: 0.0, d: 0.0}
-lhand_controller:
-#  type: effort_controllers/JointTrajectoryController
-  type: velocity_controllers/JointTrajectoryController
-  joints:
-    - l_thumb_joint
-  gains:
-    l_thumb_joint:     {p: 25.0, i: 0.0, d: 0.0}
-rhand_controller:
-#  type: effort_controllers/JointTrajectoryController
-  type: velocity_controllers/JointTrajectoryController
-  joints:
-    - r_thumb_joint
-  gains:
-    r_thumb_joint:     {p: 25.0, i: 0.0, d: 0.0}
-
-robot_interface_controllers: ['rarm', 'larm', 'waist', 'lifter', 'head']
