--- aero_description/typeF/urdf/aero-wheels.urdf.xacro
@@ -1,44 +1,0 @@
-<?xml version="1.0"?>
-
-<!-- XML namespaces -->
-<robot xmlns:xi="http://www.w3.org/2001/XInclude"
-       xmlns:gazebo="http://playerstage.sourceforge.net/gazebo/xmlschema/#gz"
-       xmlns:model="http://playerstage.sourceforge.net/gazebo/xmlschema/#model"
-       xmlns:sensor="http://playerstage.sourceforge.net/gazebo/xmlschema/#sensor"
-       xmlns:body="http://playerstage.sourceforge.net/gazebo/xmlschema/#body"
-       xmlns:geom="http://playerstage.sourceforge.net/gazebo/xmlschema/#geom"
-       xmlns:joint="http://playerstage.sourceforge.net/gazebo/xmlschema/#joint"
-       xmlns:interface="http://playerstage.sourceforge.net/gazebo/xmlschema/#interface"
-       xmlns:rendering="http://playerstage.sourceforge.net/gazebo/xmlschema/#rendering"
-       xmlns:renderable="http://playerstage.sourceforge.net/gazebo/xmlschema/#renderable"
-       xmlns:controller="http://playerstage.sourceforge.net/gazebo/xmlschema/#controller"
-       xmlns:physics="http://playerstage.sourceforge.net/gazebo/xmlschema/#physics"
-       xmlns:xacro="http://ros.org/wiki/xacro">
-
-  <xacro:include filename="$(find aero_description)/urdf/common.xacro" />
-  <xacro:include filename="$(find aero_description)/urdf/typeC_lifter/lifter.urdf.xacro" />
-  <xacro:include filename="$(find aero_description)/urdf/typeC_lifter/virtual_lifter.urdf.xacro" />
-  <xacro:include filename="$(find aero_description)/urdf/typeE_wheel/wheels.urdf.xacro" />
-  
-
-  <xacro:macro name="aero-wheels" params="base_link virtual_lifter:=false x_max:=0.0 x_min:=0.0 z_max:=0.0 z_min:=0.0 disable_mesh:=false">
-
-    <xacro:wheels base_link="${base_link}" disable_mesh="${disable_mesh}"/>
-
-    <joint name="wheel_lifter_connector" type="fixed">
-      <parent  link="wheels_top_link"/>
-      <child link="lifter_base_link"/>
-      <origin xyz="0.0 0.0 0.005" rpy="0 0 0 "/> 
-   </joint>
-
-   <xacro:unless value="${virtual_lifter}">
-    <xacro:lifter base_link="lifter_base_link"/>
-   </xacro:unless>
-
-   <xacro:if value="${virtual_lifter}">
-    <xacro:virtual_lifter base_link="lifter_base_link" x_max="${x_max}" x_min="${x_min}" z_max="${z_max}" z_min="${z_min}"/>
-   </xacro:if>
-
-  </xacro:macro>
-
-</robot>
