--- aero_description/typeF/urdf/common.xacro
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<robot>
-
-  <property name="M_PI" value="3.1415926535897931" />
-
-</robot>
