+++ aero_description/typeFCET/README.md
@@ -1,0 +1,15 @@
+# Aero Robot Setting
+
+## type FCDT
+
+- upper
+  -  typeF_upperbody
+
+- lifter
+  - typeC_lifter
+
+- wheel
+  - typeD_wheel
+
+- hand
+  - trx_s(T)
