+++ aero_description/typeFCET/aero-wheels.txt
@@ -1,0 +1,2 @@
+knee_joint
+ankle_joint
