+++ aero_description/typeFCET/moveit_config/config/fake_controllers.yaml
@@ -1,0 +1,216 @@
+controller_list:
+  - name: fake_rarm_controller
+    joints:
+      - r_shoulder_p_joint
+      - r_shoulder_r_joint
+      - r_shoulder_y_joint
+      - r_elbow_joint
+      - r_wrist_y_joint
+      - r_wrist_p_joint
+      - r_wrist_r_joint
+  - name: fake_larm_controller
+    joints:
+      - l_shoulder_p_joint
+      - l_shoulder_r_joint
+      - l_shoulder_y_joint
+      - l_elbow_joint
+      - l_wrist_y_joint
+      - l_wrist_p_joint
+      - l_wrist_r_joint
+  - name: fake_both_arms_controller
+    joints:
+      - l_shoulder_p_joint
+      - l_shoulder_r_joint
+      - l_shoulder_y_joint
+      - l_elbow_joint
+      - l_wrist_y_joint
+      - l_wrist_p_joint
+      - l_wrist_r_joint
+      - r_shoulder_p_joint
+      - r_shoulder_r_joint
+      - r_shoulder_y_joint
+      - r_elbow_joint
+      - r_wrist_y_joint
+      - r_wrist_p_joint
+      - r_wrist_r_joint
+  - name: fake_waist_controller
+    joints:
+      - waist_y_joint
+      - waist_p_joint
+      - waist_r_joint
+  - name: fake_lifter_controller
+    joints:
+      - ankle_joint
+      - knee_joint
+  - name: fake_rarm_with_waist_controller
+    joints:
+      - waist_y_joint
+      - waist_p_joint
+      - waist_r_joint
+      - r_shoulder_p_joint
+      - r_shoulder_r_joint
+      - r_shoulder_y_joint
+      - r_elbow_joint
+      - r_wrist_y_joint
+      - r_wrist_p_joint
+      - r_wrist_r_joint
+  - name: fake_larm_with_waist_controller
+    joints:
+      - waist_y_joint
+      - waist_p_joint
+      - waist_r_joint
+      - l_shoulder_p_joint
+      - l_shoulder_r_joint
+      - l_shoulder_y_joint
+      - l_elbow_joint
+      - l_wrist_y_joint
+      - l_wrist_p_joint
+      - l_wrist_r_joint
+  - name: fake_both_arms_with_waist_controller
+    joints:
+      - waist_y_joint
+      - waist_p_joint
+      - waist_r_joint
+      - l_shoulder_p_joint
+      - l_shoulder_r_joint
+      - l_shoulder_y_joint
+      - l_elbow_joint
+      - l_wrist_y_joint
+      - l_wrist_p_joint
+      - l_wrist_r_joint
+      - r_shoulder_p_joint
+      - r_shoulder_r_joint
+      - r_shoulder_y_joint
+      - r_elbow_joint
+      - r_wrist_y_joint
+      - r_wrist_p_joint
+      - r_wrist_r_joint
+  - name: fake_rarm_with_torso_controller
+    joints:
+      - ankle_joint
+      - knee_joint
+      - waist_y_joint
+      - waist_p_joint
+      - waist_r_joint
+      - r_shoulder_p_joint
+      - r_shoulder_r_joint
+      - r_shoulder_y_joint
+      - r_elbow_joint
+      - r_wrist_y_joint
+      - r_wrist_p_joint
+      - r_wrist_r_joint
+  - name: fake_larm_with_torso_controller
+    joints:
+      - ankle_joint
+      - knee_joint
+      - waist_y_joint
+      - waist_p_joint
+      - waist_r_joint
+      - l_shoulder_p_joint
+      - l_shoulder_r_joint
+      - l_shoulder_y_joint
+      - l_elbow_joint
+      - l_wrist_y_joint
+      - l_wrist_p_joint
+      - l_wrist_r_joint
+  - name: fake_rarm_with_lifter_controller
+    joints:
+      - ankle_joint
+      - knee_joint
+      - r_shoulder_p_joint
+      - r_shoulder_r_joint
+      - r_shoulder_y_joint
+      - r_elbow_joint
+      - r_wrist_y_joint
+      - r_wrist_p_joint
+      - r_wrist_r_joint
+  - name: fake_larm_with_lifter_controller
+    joints:
+      - ankle_joint
+      - knee_joint
+      - l_shoulder_p_joint
+      - l_shoulder_r_joint
+      - l_shoulder_y_joint
+      - l_elbow_joint
+      - l_wrist_y_joint
+      - l_wrist_p_joint
+      - l_wrist_r_joint
+  - name: fake_head_controller
+    joints:
+      - neck_y_joint
+      - neck_p_joint
+      - neck_r_joint
+  - name: fake_lifter_ee_controller
+    joints:
+      []
+  - name: fake_waist_ee_controller
+    joints:
+      - waist_r_joint
+  - name: fake_head_ee_controller
+    joints:
+      - neck_r_joint
+  - name: fake_rhand_controller
+    joints:
+      - r_indexbase_joint
+      - r_indexend_joint
+      - r_indexmid_joint
+      - r_thumb_joint
+  - name: fake_lhand_controller
+    joints:
+      - l_indexbase_joint
+      - l_indexend_joint
+      - l_indexmid_joint
+      - l_thumb_joint
+  - name: fake_torso_controller
+    joints:
+      - ankle_joint
+      - knee_joint
+      - waist_y_joint
+      - waist_p_joint
+      - waist_r_joint
+  - name: fake_upper_body_controller
+    joints:
+      - waist_y_joint
+      - waist_p_joint
+      - waist_r_joint
+      - l_shoulder_p_joint
+      - l_shoulder_r_joint
+      - l_shoulder_y_joint
+      - l_elbow_joint
+      - l_wrist_y_joint
+      - l_wrist_p_joint
+      - l_wrist_r_joint
+      - neck_y_joint
+      - neck_p_joint
+      - neck_r_joint
+      - r_shoulder_p_joint
+      - r_shoulder_r_joint
+      - r_shoulder_y_joint
+      - r_elbow_joint
+      - r_wrist_y_joint
+      - r_wrist_p_joint
+      - r_wrist_r_joint
+  - name: fake_whole_body_controller
+    joints:
+      - ankle_joint
+      - knee_joint
+      - waist_y_joint
+      - waist_p_joint
+      - waist_r_joint
+      - l_shoulder_p_joint
+      - l_shoulder_r_joint
+      - l_shoulder_y_joint
+      - l_elbow_joint
+      - l_wrist_y_joint
+      - l_wrist_p_joint
+      - l_wrist_r_joint
+      - neck_y_joint
+      - neck_p_joint
+      - neck_r_joint
+      - r_shoulder_p_joint
+      - r_shoulder_r_joint
+      - r_shoulder_y_joint
+      - r_elbow_joint
+      - r_wrist_y_joint
+      - r_wrist_p_joint
+      - r_wrist_r_joint
